requirements.txt
# ---------- Core Backend ----------
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ---------- Frontend ----------
streamlit==1.32.2
requests==2.31.0

# ---------- LLM ----------
openai==0.28.1

# ---------- Utilities ----------
python-dotenv==1.0.1
pydantic==2.6.4

# ---------- External Data ----------
wikipedia==1.4.0
